+++ src/Styles/Player.css
@@ -13,10 +13,7 @@
     border-radius: 40vw;
     
 }
-.image-col{
 
-    
-}
 .player-col{
     display: flex;
     align-items: end;
@@ -53,7 +50,8 @@
 
 }
 .control{
-    margin-left: 5vw;
+    margin-left: 2%;
+    margin-right: 2%;
     font-size: 40px;
     color: #000000;
     
@@ -66,6 +64,69 @@
 }
 .duration{
     display:flex;
-    width: 90%;
     justify-content: space-between;
+    
+    margin-bottom: 1%;
+}
+
+.bottom-player{
+    background-color: #FFED00 !important;
+    max-height: 10vh !important;
+    
+}
+.bottom-controls{
+
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    
+}
+.bottom-player-image{
+
+    max-height: 90%;
+    max-width:90%;
+    max-height: 10vh;
+    
+    
+}
+.bottom-player-seekbar{
+
+    display:inline-block;
+    width:65%;
+    height: 1px;
+    
+}
+.bottom-progessbar{
+
+    display: block;
+    
+    width: 100%;
+    
+    height: 3px; /* Specified height */
+    background: #000000; /* Grey background */
+    outline: none; /* Remove outline */
+    opacity: 1; /* Set transparency (for mouse-over effects on hover) */
+    -webkit-transition: .2s; /* 0.2 seconds transition on hover */
+    transition: opacity .2s;
+    
+}
+input[type=range]{
+    -webkit-appearance: none;
+}
+input[type=range]::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    border: none;
+    height: 16px;
+    width: 16px;
+    border-radius: 50%;
+    background: rgb(255, 255, 255);
+    margin-top: -4px;
+}
+.card-title{
+    font-family: 'Bebas Neue', cursive;
+    letter-spacing: 1px;
+}
+.card-text{
+
+    font-family: 'Gidole', sans-serif;
 }
